% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{baselineEs}
\alias{baselineEs}
\title{baselineEs}
\usage{
baselineEs(chrDf, threshold = 1, tol_m = 30, loops = 6)
}
\arguments{
\item{chrDf}{A chrDf.}

\item{threshold}{threshold for inflect.}

\item{tol_m}{Tolerable angles.}

\item{loops}{loop number.}
}
\value{
A new chrDf with baseline column.
}
\description{
Estimating the baseline for a chrDf.
}
\examples{
chrDf_j <- ZOIList[[j]]
chrDf_j <- baselineEs(chrDf_j, loops = 8, tol_m = 30)
}
