% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{bins2ZOIs}
\alias{bins2ZOIs}
\title{bins2ZOIs}
\usage{
bins2ZOIs(
  chrDfList,
  smooth = "mean",
  size = 3,
  p = 3,
  etlD = 1,
  IETH = 1.75,
  preNum = 3,
  sn = 3,
  loops = 8,
  tol_m = 30,
  threshold = 1,
  thread = 1
)
}
\arguments{
\item{chrDfList}{A chrDfList, one chrDf means a bin.}

\item{smooth}{smooth type.}

\item{size}{smoothMean size.}

\item{p}{smoothSg p.}
}
\value{
A chrDfList.
}
\description{
Bins 's chrDf to ZOIs's chrDf.
}
\examples{
bins2ZOIs(chrDfList = chrDf_bins, thread = 4)
}
