% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{cal_IE}
\alias{cal_IE}
\title{cal_IE}
\usage{
cal_IE(int)
}
\arguments{
\item{int}{int.}
}
\value{
IE.
}
\description{
Calculate IE.
}
\examples{
intensity <- c(1701.196, 2121.159, 2401.849, 2043.653, 1542.964,  723.803)
cal_IE(int = intensity)
}
