% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{checkCentroid}
\alias{checkCentroid}
\title{checkCentroid}
\usage{
checkCentroid(file_path)
}
\arguments{
\item{file_path}{A vector of file's path.}
}
\value{
A logical vector.
}
\description{
Check if the input file is centroid mode.
}
\examples{
file_dir <- "D:/fudan/Projects/2024/MetaboProcess/Progress/build_package/test_data/massProcesser/"
file_path <- list.files(file_dir, pattern = ".mzXML")
file_path <- paste0(file_dir, file_path)
checkCentroid(file_path = file_path)
}
