% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{FindIso}
\alias{FindIso}
\title{FindIso}
\usage{
FindIso(ZOITable, data_QC, thread = 3)
}
\arguments{
\item{ZOITable}{ZOITable.}

\item{data_QC}{data_QC.}

\item{thread}{thread.}
}
\value{
A ISOTable.
}
\description{
Find isotopes from ZOI.
}
\examples{
isoTable <- FindIso(ZOITable, data_QC, thread = 3)
}
