% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{FindZOI}
\alias{FindZOI}
\title{FindZOI}
\usage{
FindZOI(chrDf, noise = NA, preNum = 3, etlD = 1, IETH = 1.75)
}
\arguments{
\item{chrDf}{A chrDf.}

\item{noise}{noise.}

\item{preNum}{preNum.}

\item{etlD}{etlD. Extend width.}

\item{IETH}{IE threshold.}
}
\value{
A chrDfList.
}
\description{
Find ZOIs from a chrDf of the bin.
}
\examples{
ZOIList <- FindZOI(chrDf = chrDf, noise = noiseEstimation(chrDf), etlD = 1, IETH = 10)
}
