% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{FindZOI_baseline}
\alias{FindZOI_baseline}
\title{FindZOI_baseline}
\usage{
FindZOI_baseline(chrDf, noise, sn = 0, preNum = 3)
}
\arguments{
\item{chrDf}{chrDf}

\item{noise}{noise}

\item{sn}{sn}

\item{preNum}{preNum}
}
\value{
A chrDf list
}
\description{
Find ZOIs from a big ZOI by baseline.
}
\examples{
FindZOI_baseline(chrDf, noise = 1000)
}
