% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{generateBin_old}
\alias{generateBin_old}
\title{generateBin_old}
\usage{
generateBin_old(ndata, bin = 0.05, mslevel = 1, thread = 1)
}
\arguments{
\item{ndata}{nth sample data, it is a MsExperiment object.}

\item{bin}{bin mz.}

\item{mslevel}{mslevel, 1 or 2.}

\item{thread}{Number of threads in parallel.}
}
\value{
A list consisting of chrDf.
}
\description{
Divide a sample into multiple chromatograms according to the length of the bin.
}
\examples{
chrDf_bins <- generateBin(ndata = ndata, bin = 0.05, mslevel = 1, thread = 24)
}
