% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_baselinePara}
\alias{get_baselinePara}
\title{get_baselinePara}
\usage{
get_baselinePara(threshold = 1, tol_m = 30, loops = 6)
}
\arguments{
\item{threshold}{threshold.}

\item{tol_m}{tol_m.}

\item{loops}{loops.}
}
\value{
A parameters list.
}
\description{
Get baselineEs function parameters.
}
\examples{
baselinePara <- getbaselinePara()
}
