% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_smoothPara}
\alias{get_smoothPara}
\title{get_smoothPara}
\usage{
get_smoothPara(
  smooth = TRUE,
  method = "mean",
  size = 3,
  p = 3,
  n = p + 3 - p\%\%2,
  m = 0,
  ts = 1
)
}
\arguments{
\item{smooth}{Whether to perform smooth.}

\item{method}{Smooth method, sg or mean.}

\item{size}{SmoothMean size.}

\item{p}{sg p.}

\item{n}{sg n.}

\item{m}{sg m.}

\item{ts}{sg ts.}
}
\value{
A parameter list.
}
\description{
Get smooth parameters list.
}
\examples{
smoothPara <- get_smoothPara()
}
