% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{getChromPeaksDf}
\alias{getChromPeaksDf}
\title{getChromPeaksDf}
\usage{
getChromPeaksDf(
  data,
  cpid = NA,
  noise1 = 0,
  noise2 = 0,
  smoothPara = get_smoothPara(),
  expandRt = 0,
  expandMz = 0
)
}
\arguments{
\item{data}{A XcmsExperiment object.}

\item{cpid}{cpid vector.}

\item{noise1}{noise for ms1.}

\item{noise2}{noise for ms2.}

\item{smoothPara}{smoothPara.}

\item{expandRt}{expandRt.}

\item{expandMz}{expandMz}
}
\value{
A chrDfList.
}
\description{
Get chromPeak date frame.
}
\examples{
getChromPeaksDf(data)
}
