% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{getChromPeakTable}
\alias{getChromPeakTable}
\title{getChromPeakTable}
\usage{
getChromPeakTable(data, style = "xcms")
}
\arguments{
\item{data}{A xcmsExperiment object after peak picking.}

\item{style}{c("xcms", "neatms")}
}
\value{
A tibble.
}
\description{
Get a chromPeakTable tibble.
}
\examples{
getChromPeakTable(data)
}
