% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{getFeatureTable}
\alias{getFeatureTable}
\title{getFeatureTable}
\usage{
getFeatureTable(data, method = "medret", value = "intb")
}
\arguments{
\item{data}{A XcmsExperiment object.}

\item{method}{See xcms::featureValues function.}

\item{value}{See xcms::featureValues function.}
}
\value{
A tibble.
}
\description{
Get feature table.
}
\examples{
fearureTable <- getFeatureTable(data)
}
