% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{noiseEstimation}
\alias{noiseEstimation}
\title{noiseEstimation}
\usage{
noiseEstimation(chrDf)
}
\arguments{
\item{chrDf}{A chrDf.}
}
\value{
A numeric.
}
\description{
Estimates the noise of a chromatogram, which can be used to find the apex of a chromatographic peak.
}
\examples{
noiseEstimation(chrDf_bins[[30]])
}
