% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{optParam4xcms}
\alias{optParam4xcms}
\title{optParam4xcms_old}
\usage{
optParam4xcms()
}
\arguments{
\item{data_QC}{A MsExperiment object which contains QC sample.}

\item{res_dir}{Results path.}

\item{bin}{bin size.}

\item{mslevel}{mslevel.}

\item{thread1}{thread1.}

\item{output_bins}{Whether to output bins.}

\item{bin_scanNum}{bin's scanNum threshold.}

\item{smooth}{smooth methods, c("mean", "sg").}

\item{size}{mean size.}

\item{p}{p.}

\item{etlD}{etlD.}

\item{IETH}{IETH.}

\item{preNum}{preNum.}

\item{sn}{sn.}

\item{loops}{loops.}

\item{tol_m1}{tol_m1.}

\item{thread2}{thread2.}

\item{threshold}{threshold.}

\item{tol_m2}{tol_m2}

\item{output_ZOI}{Whether to output ZOIs.}

\item{output_ZOI2}{Whether to output ZOI2s.}

\item{factor}{factor.}

\item{range}{range.}

\item{massRange}{massRange.}

\item{output_ZOI3}{Whether to output ZOI3s.}

\item{rt_tol}{rt_tol.}

\item{mz_tol}{mz_tol.}

\item{tol_nf}{tol_nf.}

\item{method}{method.}

\item{shiftRange}{shiftRange.}

\item{widthRange}{widthRange.}

\item{maxMassTol}{maxMassTol.}

\item{minPeakWidth}{minPeakWidth.}

\item{maxPeakWidth}{maxPeakWidth.}
}
\value{
A list contains optimal parameters and ggplot2 object.
}
\description{
Find the optimal xcms parameters
}
\examples{
optParam4xcms(data_QC)
}
