% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paramounter_function.R
\name{paramounterPart1}
\alias{paramounterPart1}
\title{paramounterPart1}
\usage{
paramounterPart1(
  file_path,
  massSDrange = 2,
  smooth = 0,
  cutoff = 0.95,
  thread = 1,
  msLevel = 1L
)
}
\arguments{
\item{file_path}{Files path of QC sample.}

\item{massSDrange}{massSDrange.}

\item{smooth}{smooth.}

\item{cutoff}{cutoff.}

\item{thread}{thread.}

\item{msLevel}{msLevel.}
}
\value{
A number - ppmCut.
}
\description{
Paramounter part1 to calculate ppmCut.
}
\examples{
ppmCut <- paramounterPart1(file_path = file_path, thread = 24)
}
