% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{peakAnnotation}
\alias{peakAnnotation}
\title{peakAnnotation.}
\usage{
peakAnnotation(data, polarity = "positive", adinfo, thread = 1)
}
\arguments{
\item{data}{A XcmsExperiment object.}

\item{polarity}{positive or negative}

\item{adinfo}{adinfo <- data(positive.adinfo, package = "cliqueMS")}

\item{thread}{thread.}
}
\value{
A tibble.
}
\description{
peakAnnotation.
}
\examples{
data(positive.adinfo, package = "cliqueMS")
positive.adinfo <- positive.adinfo[positive.adinfo$adduct \%in\%
                                     c("[M+H]+", "[M+H-H2O]+", "[M+Na]+", "[M+H-NH3]+", "[M+K]+", "[M+NH4]+"),]
resTable <- peakAnnotation(data = data_filter, polarity = "positive", adinfo = positive.adinfo, thread = length(data_filter))
}
