% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{pickZOI}
\alias{pickZOI}
\title{pickZOI}
\usage{
pickZOI(
  chrDf,
  smoothPara = get_smoothPara(),
  baselinePara = get_baselinePara(),
  sn = 1,
  preNum = 3,
  tol_m = 10,
  snthresh = 0.5
)
}
\arguments{
\item{chrDf}{A chrDf.}

\item{smoothPara}{smoothPara.}

\item{baselinePara}{baselinePara.}

\item{sn}{For filtering peaks with low signal-to-noise ratio.}

\item{preNum}{A peak needs to have preNum of data points greater than noise0.}

\item{tol_m}{Angle tolerance for baseline calculation and edge finding.}

\item{snthresh}{The signal-to-noise ratio used for the MatchedFilter,
which hides the peak when the signal-to-noise ratio is lower than this value.}
}
\value{
A ZOI list contains chrDf.
}
\description{
Direct peak(ZOI) finding on one chromatogram.
}
\examples{
data("chrDf_mrm_test")
plot_chrDf(chrDf_mrm_test)
ZOIList <- pickZOI(chrDf = chrDf_mrm_test, sn = 3)
length(ZOIList) # 6
plot_chrDf(ZOIList[[1]], baseline = TRUE)
}
