% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{smoothMean}
\alias{smoothMean}
\title{smoothMean}
\usage{
smoothMean(int, size = 3)
}
\arguments{
\item{int}{intensity.}

\item{size}{mean size.}
}
\value{
intensity vector.
}
\description{
Smooth function using mean.
}
\examples{
intensity <- c(1701.196, 2121.159, 2401.849, 2043.653, 1542.964,  723.803)
smoothMean(int = intensity)
}
