% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{smoothSg}
\alias{smoothSg}
\title{smoothSg}
\usage{
smoothSg(int, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)
}
\arguments{
\item{int}{intensity.}

\item{p}{filter order.}

\item{n}{filter length (must be odd).}

\item{m}{return the m-th derivative of the filter coefficients.}

\item{ts}{time scaling factor.}
}
\value{
intensity vector
}
\description{
Smooth data with a Savitzky-Golay smoothing filter.
}
\examples{
intensity <- c(1701.196, 2121.159, 2401.849, 2043.653, 1542.964,  723.803)
smoothSg(int = intensity)
}
