% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{ZOIList2ZOITable}
\alias{ZOIList2ZOITable}
\title{ZOIList2ZOITable}
\usage{
ZOIList2ZOITable(ZOIList)
}
\arguments{
\item{ZOIList}{ZOIList}
}
\value{
ZOITable.
}
\description{
Make ZZOIList to ZOITable.
}
\examples{
ZOITable.
ZOIList <- readRDS("D:/fudan/Projects/2024/MetaboProcess/Progress/generate_bins/240705/ZOIList.rds")
ZOITable <- ZOIList2ZOITable(ZOIList)
}
