% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{generateBin}
\alias{generateBin}
\title{generateBin}
\usage{
generateBin(ndata, bin = 0.05, slide = 0.05, mslevel = 1, thread = 1)
}
\arguments{
\item{ndata}{nth sample data, it is a MsExperiment object.}

\item{bin}{bin mz.}

\item{slide}{slide size.}

\item{mslevel}{mslevel.}

\item{thread}{thread.}
}
\value{
A chrDf list.
}
\description{
Divide a sample into multiple chromatograms according to the length of the bin.
}
\examples{
chrDfList_bins <- generateBin(ndata = ndata, thread = 4)
}
