% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{optParam4xcms}
\alias{optParam4xcms}
\title{optParam4xcms}
\usage{
optParam4xcms(
  data_QC,
  res_dir = "./",
  bin = 0.05,
  slide = 0.05,
  mslevel = 1,
  thread1 = 1,
  output_bins = FALSE,
  bin_scanNum = 100,
  smoothPara = get_smoothPara(),
  baselinePara = get_baselinePara(),
  sn = 3,
  preNum = 3,
  tol_m = 10,
  snthresh = 0.5,
  thread2 = 1,
  maxMassTol = 100,
  minPeakWidth = 3,
  maxPeakWidth = 100,
  thread3 = 1
)
}
\arguments{
\item{data_QC}{A MsExperiment object which contains QC sample.}

\item{res_dir}{Results path.}

\item{bin}{bin size.}

\item{slide}{slide size.}

\item{mslevel}{mslevel.}

\item{thread1}{thread1.}

\item{output_bins}{Whether to output bins.}

\item{bin_scanNum}{bin's scanNum threshold.}

\item{smoothPara}{smoothPrar.}

\item{baselinePara}{baselinePara}

\item{sn}{sn}

\item{preNum}{preNum.}

\item{tol_m}{tol_m.}

\item{snthresh}{snthresh.}

\item{thread2}{thread2.}
}
\value{
A parameter vector.
}
\description{
optParam4xcms.
}
\examples{
file_dir <- "D:/fudan/Projects/2024/MetaboProcess/Data/SRM 1950 MassIVE MSV000083469/"
patterns <- c(".mzXML", ".mzxml", ".mzML", ".mzml")
patterns <- paste0(patterns, collapse = "|")
file_path <- list.files(file_dir, pattern = patterns)
file_path <- paste0(file_dir, file_path)
pd <- data.frame(sample_name = sub(basename(file_path), pattern = patterns,
                                   replacement = ""),
                 sample_group = rep("QC", 5),
                 sample_type = rep("QC", 5),
                 sample_inject = c(1, 3, 4, 5, 2),
                 sample_path = file_path,
                 stringsAsFactors = FALSE)
pd <- dplyr::arrange(pd, sample_inject)
data <- MsExperiment::readMsExperiment(pd$sample_path, sampleData = pd)
data_QC <- data
optParam4xcms(data_QC = data_QC, thread2 = 4)
}
