% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_parameter_optimization.R
\name{optParam4xcms}
\alias{optParam4xcms}
\title{optParam4xcms}
\usage{
optParam4xcms(
  data_QC,
  res_dir = "./",
  bin = 0.05,
  mslevel = 1,
  thread1 = 1,
  output_bins = FALSE,
  bin_scanNum = 10,
  smooth = "mean",
  size = 3,
  p = 3,
  etlD = 1,
  IETH = 10,
  preNum = 3,
  sn = 0,
  loops = 8,
  tol_m1 = 30,
  threshold = 1,
  thread2 = 1,
  tol_m2 = 40,
  output_ZOI = FALSE,
  output_ZOI2 = FALSE,
  factor = 1,
  range = 1,
  massRange = 0.99,
  output_ZOI3 = FALSE,
  rt_tol = 5,
  mz_tol = 0.02,
  tol_nf = 0.5,
  method = "mz",
  widthRange = 0.99,
  shiftRange = 0.95,
  maxMassTol = 100,
  minPeakWidth = 1,
  maxPeakWidth = 100
)
}
\arguments{
\item{data_QC}{A MsExperiment object which contains QC sample.}

\item{res_dir}{Results path.}

\item{bin}{bin size.}

\item{mslevel}{mslevel.}

\item{thread1}{thread1.}

\item{output_bins}{Whether to output bins.}

\item{bin_scanNum}{bin's scanNum threshold.}

\item{smooth}{smooth methods, c("mean", "sg").}

\item{size}{mean size.}

\item{p}{p.}

\item{etlD}{etlD.}

\item{IETH}{IETH.}

\item{preNum}{preNum.}

\item{sn}{sn.}

\item{loops}{loops.}

\item{tol_m1}{tol_m1.}

\item{threshold}{threshold.}

\item{thread2}{thread2.}

\item{tol_m2}{tol_m2}

\item{output_ZOI}{Whether to output ZOIs.}

\item{output_ZOI2}{Whether to output ZOI2s.}

\item{factor}{factor.}

\item{range}{range.}

\item{massRange}{massRange.}

\item{output_ZOI3}{Whether to output ZOI3s.}

\item{rt_tol}{rt_tol.}

\item{mz_tol}{mz_tol.}

\item{tol_nf}{tol_nf.}

\item{method}{method.}

\item{widthRange}{widthRange.}

\item{shiftRange}{shiftRange.}

\item{maxMassTol}{maxMassTol.}

\item{minPeakWidth}{minPeakWidth.}

\item{maxPeakWidth}{maxPeakWidth.}
}
\value{
A list contains optimal parameters and ggplot2 object.
}
\description{
Find the optimal xcms parameters
}
\examples{
optParam4xcms(data_QC)
}
